nvme: mi: Improve exploitation of deku for request parsing
Primarily, parse subtype bodies using enums, handling split
discriminants with deku's ctx and id attributes.

The result of this is that, rather than producing a series of subtype
fragments in unrelated structs, we have one struct instance that
contains the parse for any subtype(s). With that, we can now pass that
one struct instance as context to the RequestHandler trait as we walk
the subtype descriminants. This, in-turn, enables inspection of context
that is unavailable in the current subtype struct, such as extraction of
CTLID for the Admin command collection.

Response assembly is still done piece-meal. This may be improved later,
but the ergonomics aren't too bad given vectored send.

Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>